requirements.txt
annotated-types==0.6.0
anyio==4.0.0
ariadne==0.21
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
graphql-core==3.2.3
greenlet==3.0.1
idna==3.4
inflect==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
sniffio==1.3.0
sqlacodegen==3.0.0rc3
SQLAlchemy==1.4.50
starlette==0.32.0.post1
typing_extensions==4.8.0
Werkzeug==3.0.1
